+++ src/tquality_selenium/elements/__init__.py
@@ -1,6 +1,7 @@
 from tquality_selenium.elements.base_element import BaseElement
 from tquality_selenium.elements.button import Button
-from tquality_selenium.elements.by import By, ByKind
+from tquality_selenium.elements.by import By
+from tquality_selenium.elements.by_kind import ByKind
 from tquality_selenium.elements.checkbox import CheckBox
 from tquality_selenium.elements.input import Input
 from tquality_selenium.elements.label import Label