+++ src/tquality_selenium/services/__init__.py
@@ -4,7 +4,6 @@ from tquality_selenium.services.collection_factory import (
 )
 from tquality_selenium.services.element_factory import ElementFactory
 from tquality_selenium.services.lazy_elements import LazyElements
-from tquality_selenium.services.element_waiter import ElementWaiter
 from tquality_selenium.services.js_actions import (
     ElementJsActions,
     JsActions,
@@ -18,7 +17,6 @@ __all__ = [
     "DomField",
     "ElementFactory",
     "ElementJsActions",
-    "ElementWaiter",
     "JsActions",
     "LazyElements",
     "PseudoElement",